distances: Refactor into separate files
